Add files via upload
The fact that ATR is calculated using absolute values of differences in price is something that should not be ignored. This is relevant because it means that securities with higher price values will inherently have higher ATR values. Likewise, securities with lower price values will have lower ATR values. The consequence is that a trader cannot compare the ATR Values of multiple securities. What is considered to be a high ATR Value or a high ATR Range for one security may not be the same for another security. A trader should study and research the relevance of ATR for each security independently when performing chart analysis.